requirements.txt
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
openai>=1.30.0
aiosqlite>=0.20.0
lxml>=5.0.0
playwright>=1.40.0
